Change `dbms.procedures` to be a `DBMS` procedure
`dbms.procedures` had been implemented as a `READ` procedure,
which was requiring read transactions on the database. This is not
necessary for inspecting loaded procedures.
