+++ community/kernel/src/main/java/org/neo4j/kernel/builtinprocs/BuiltInProcedures.java
@@ -184,19 +184,6 @@ public class BuiltInProcedures
                 .onClose( statement::close );
     }
 
-    @Description( "List all procedures in the DBMS." )
-    @Procedure( name = "dbms.procedures", mode = READ )
-    public Stream<ProcedureResult> listProcedures()
-    {
-        try ( Statement statement = tx.acquireStatement() )
-        {
-            return statement.readOperations().proceduresGetAll()
-                    .stream()
-                    .sorted( ( a, b ) -> a.name().toString().compareTo( b.name().toString() ) )
-                    .map( ProcedureResult::new );
-        }
-    }
-
     @Description( "List all user functions in the DBMS." )
     @Procedure(name = "dbms.functions", mode = READ)
     public Stream<FunctionResult> listFunctions()
@@ -271,20 +258,6 @@ public class BuiltInProcedures
         private ConstraintResult( String description )
         {
             this.description = description;
-        }
-    }
-
-    public class ProcedureResult
-    {
-        public final String name;
-        public final String signature;
-        public final String description;
-
-        private ProcedureResult( ProcedureSignature signature )
-        {
-            this.name = signature.name().toString();
-            this.signature = signature.toString();
-            this.description = signature.description().orElse( "" );
         }
     }
